+++ apps/RNApp/brownfield.config.js
@@ -1,0 +1,13 @@
+/**
+ * @type {import('@callstack/react-native-brownfield').BrownfieldConfig}
+ */
+module.exports = {
+  android: {
+    moduleName: ':BrownfieldLib',
+    varian: 'devRelease',
+  },
+  ios: {
+    scheme: 'BrownfieldLib',
+  },
+  verbose: true,
+};