+++ packages/cli/src/brownie/index.ts
@@ -2,7 +2,6 @@ import { styleText } from 'node:util';
 
 import * as Commands from './commands/index.js';
 
-export type * from './types.js';
 export * from './store-discovery.js';
 
 export const groupName = `${styleText(['bold', 'blueBright'], '@callstack/brownie')}${styleText('whiteBright', ' - Shared state management CLI for React Native Brownfield')}`;