+++ packages/cli/src/brownie/types.ts
@@ -1,1 +1,0 @@
-export type Platform = 'swift' | 'kotlin';